🧃 api | added recurring cleaning of data to reduce unexpected data loss

+++ app/DataIntegrators/FalkRossHandler.php
@@ -93,11 +93,14 @@ class FalkRossHandler extends ApiHandler
             }
 
             $this->sync->addLog("in progress (step)", 2, "Product downloaded", (++$counter / $total) * 100);
+
+            $started_at ??= now();
+            if ($started_at < now()->subMinutes(1)) {
+                if ($this->sync->product_import_enabled) $this->deleteUnsyncedProducts($this->imported_ids);
+                $started_at = now();
+            }
         }
 
-        if ($this->sync->product_import_enabled) {
-            $this->deleteUnsyncedProducts($this->imported_ids);
-        }
         $this->reportSynchCount($counter, $total);
     }
     #endregion
